src/model/requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1623774798911/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1623682282434/work
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1610094717693/work
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
blinker==1.4
brotlipy==0.7.0
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1619593780156/work
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1625835338884/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454189/work
click @ file:///Users/runner/miniforge3/conda-bld/click_1621503761280/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851494399/work
cycler==0.10.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1620718971772/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1624952364672/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1625312698221/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1624380665635/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1617739685813/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1625463774392/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1625027276326/work
ipython-genutils==0.2.0
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146860797/work
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1624926174587/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1614595805172/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1620506033672/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1624981217218/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1624399904045/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1622563202229/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1617859230218/work
pandas==1.3.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1625678155585/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1623977816122/work
protobuf==3.17.2
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1620245170812/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1619620644068/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291456365/work
python-dateutil==2.8.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1595492159598/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1614171254180/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1624824918804/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1610699261066/work/tensorboard-2.4.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1611075653546/work/tensorboard_plugin_wit-1.8.0-py3-none-any.whl
tensorflow-addons @ file:///Users/luis/Downloads/tensorflow_macos/arm64/tensorflow_addons-0.11.2%2Bmlcompute-cp38-cp38-macosx_11_0_arm64.whl
tensorflow-estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-estimator_1623773901369/work/wheel_dir/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
tensorflow-macos @ file:///Users/luis/Downloads/tensorflow_macos/arm64/tensorflow_macos-0.1a1-cp38-cp38-macosx_11_0_arm64.whl
termcolor==1.1.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1625488911690/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typeguard @ file:///home/conda/feedstock_root/build_artifacts/typeguard_1622818924333/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1621518206714/work
wrapt @ file:///Users/runner/miniforge3/conda-bld/wrapt_1624971819058/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1625233166158/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1625284368454/work
